refactor: move nrf-specific linking to ariel-os-nrf
