+++ src/ariel-os-rt/Cargo.toml
@@ -17,7 +17,6 @@ linkme = { workspace = true }
 
 [build-dependencies]
 ariel-os-build = { workspace = true }
-ld-memory = { workspace = true, features = ["build-rs"], optional = true }
 
 [target.'cfg(context = "cortex-m")'.dependencies]
 cortex-m = { workspace = true }
@@ -58,8 +57,6 @@ panic-printing = []
 _panic-handler = []
 single-core = ["cortex-m/critical-section-single-core"]
 multi-core = ["embassy-rp/critical-section-impl"]
-memory-x = ["dep:ld-memory"]
-nrf91-modem = []
 
 # features needed for `cargo test`
 _test = []
