+++ src/ariel-os/Cargo.toml
@@ -178,7 +178,7 @@ ble-peripheral = ["ariel-os-embassy/ble-peripheral", "ble"]
 ble-config-static-address = ["ariel-os-embassy/ble-config-static-address"]
 
 # ## Enables support for the nRF91xx SiP modem.
-nrf91-modem = ["ariel-os-embassy/nrf91-modem", "ariel-os-rt/nrf91-modem"]
+nrf91-modem = ["ariel-os-embassy/nrf91-modem"]
 # Adjusts nRF modem (or related components) to assume that there is the DECT firmware running on the radio core.
 nrf-radiocore-firmware-dect = ["ariel-os-nrf/radiocore-firmware-dect"]
